+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.2.0 (2023-07-10)
+
+### feat: Basic implementation of smart contracts
+
+- Basic implementation of smart contracts
+  - ProjectFactory: Create an ERC-1155 NFT collection from another contract
+  - ArtistProject: The ERC-1155 NFT collection (created by ProjectFactory)
+  - TuneTogether: Main contract
+- Update deploy script
+
+---
+
 ## 0.1.0 (2023-07-07)
 
 ### feat: Init project